fix: rename and move some files
