+++ src/components/DimensionMenu/DimensionMenu.js
@@ -1,104 +1,77 @@
-import {
-    AXIS_ID_COLUMNS,
-    AXIS_ID_FILTERS,
-    AXIS_ID_ROWS,
-    VIS_TYPE_LINE_LIST,
-} from '@dhis2/analytics'
-import i18n from '@dhis2/d2-i18n'
-import { MenuDivider, FlyoutMenu, MenuItem } from '@dhis2/ui'
+import { Layer, Popper, IconMore16 } from '@dhis2/ui'
 import PropTypes from 'prop-types'
-import React from 'react'
-import { getAxisName } from '../../modules/axis.js'
+import React, { useState, useRef } from 'react'
+import { useSelector, useDispatch } from 'react-redux'
+import {
+    acAddUiLayoutDimensions,
+    acRemoveUiLayoutDimensions,
+} from '../../actions/ui.js'
+import { sGetUiType, sGetUiLayout } from '../../reducers/ui.js'
+import IconButton from '../IconButton/IconButton.js'
+import MenuItems from './MenuItems.js'
 
-const getAxisItemLabel = (axisName, isDimensionInLayout) =>
-    isDimensionInLayout
-        ? i18n.t('Move to {{axisName}}', { axisName })
-        : i18n.t('Add to {{axisName}}', { axisName })
-
-const getDividerItem = (key) => <MenuDivider dense key={key} />
-
-const DimensionMenu = ({
-    dimensionId,
-    currentAxisId,
-    visType,
-    axisItemHandler,
-    removeItemHandler,
-    onClose,
-    dataTest,
-}) => {
-    const menuItems = []
-
-    const isDimensionInLayout = !!currentAxisId
-
-    // add/move to axis item
-    const availableAxisIds =
-        visType === VIS_TYPE_LINE_LIST
-            ? [AXIS_ID_COLUMNS, AXIS_ID_FILTERS]
-            : [AXIS_ID_COLUMNS, AXIS_ID_ROWS, AXIS_ID_FILTERS]
-
-    const applicableAxisIds = availableAxisIds.filter(
-        (axisId) => axisId !== currentAxisId
+const getAxisIdForDimension = (dimensionId, layout) => {
+    const axisLayout = Object.entries(layout).find(([, dimensionIds]) =>
+        dimensionIds.includes(dimensionId)
     )
 
-    const getRemoveMenuItem = (onClick) => (
-        <MenuItem
-            key="remove-menu-item"
-            onClick={onClick}
-            label={i18n.t('Remove')}
-            dataTest={`${dataTest}-item-remove-${dimensionId}`}
-        />
-    )
+    return axisLayout ? axisLayout[0] : undefined
+}
 
-    menuItems.push(
-        ...applicableAxisIds.map((axisId) => (
-            <MenuItem
-                key={`${dimensionId}-to-${axisId}`}
-                onClick={() => {
-                    axisItemHandler({
-                        dimensionId,
-                        axisId,
-                    })
-                    onClose()
-                }}
-                label={getAxisItemLabel(
-                    getAxisName(axisId),
-                    isDimensionInLayout
-                )}
-                dataTest={`${dataTest}-item-action-${dimensionId}-to-${axisId}`}
-            />
-        ))
-    )
+const DimensionMenu = ({ currentAxisId, dimensionId }) => {
+    const dispatch = useDispatch()
+    const visType = useSelector(sGetUiType)
+    const layout = useSelector(sGetUiLayout)
 
-    // remove item
-    if (isDimensionInLayout) {
-        // divider
-        if (applicableAxisIds.length) {
-            menuItems.push(getDividerItem('remove-item-divider'))
-        }
+    const axisId = currentAxisId || getAxisIdForDimension(dimensionId, layout)
 
-        menuItems.push(
-            getRemoveMenuItem(() => {
-                removeItemHandler(dimensionId)
-                onClose()
-            })
-        )
+    const buttonRef = useRef()
+    const [menuIsOpen, setMenuIsOpen] = useState(false)
+
+    const toggleMenu = () => setMenuIsOpen(!menuIsOpen)
+
+    const getMenuId = () => `menu-for-${dimensionId}`
+
+    const axisItemHandler = ({ dimensionId, axisId }) => {
+        dispatch(acAddUiLayoutDimensions({ [dimensionId]: { axisId } }))
     }
 
+    const removeItemHandler = (id) => dispatch(acRemoveUiLayoutDimensions(id))
+
     return (
-        <FlyoutMenu dense dataTest={dataTest}>
-            {menuItems}
-        </FlyoutMenu>
+        <>
+            <div ref={buttonRef}>
+                <IconButton
+                    ariaOwns={menuIsOpen ? getMenuId() : null}
+                    ariaHaspopup={true}
+                    onClick={toggleMenu}
+                    dataTest={`layout-dimension-menu-button-${dimensionId}`}
+                >
+                    <IconMore16 />
+                </IconButton>
+            </div>
+            {menuIsOpen && (
+                <Layer onClick={toggleMenu}>
+                    <Popper reference={buttonRef} placement="bottom-start">
+                        <MenuItems
+                            dimensionId={dimensionId}
+                            currentAxisId={axisId}
+                            visType={visType}
+                            axisItemHandler={axisItemHandler}
+                            removeItemHandler={removeItemHandler}
+                            onClose={toggleMenu}
+                            dataTest={'layout-dimension-menu-dimension-menu'}
+                        />
+                    </Popper>
+                </Layer>
+            )}
+        </>
     )
 }
 
 DimensionMenu.propTypes = {
-    axisItemHandler: PropTypes.func.isRequired,
-    removeItemHandler: PropTypes.func.isRequired,
-    onClose: PropTypes.func.isRequired,
     currentAxisId: PropTypes.string,
-    dataTest: PropTypes.string,
     dimensionId: PropTypes.string,
-    visType: PropTypes.string,
 }
 
 export default DimensionMenu
